requirements.txt
openai
python-dotenv
gradio
langchain
langchain-community
langchainhub
langchain-openai
faiss-cpu